Moved command linker file to common directory.
The command linker file was duplicated in every example
project. Moved it to a common folder and added links to
each project that point to the single command linker
file. Now if any changes are made to the file they only
need to be made in one place. Can be changed back after
development if necessary.

Signed-off-by: Jason Reeder <jreeder@ti.com>