Updated the temp sensor application
Updated the temperature sensor application to use the
new bit field definitions from the updated pru_ecap.h
file.

Also, there was previously a dependency on core 1 being
loaded first that caused these examples to not be
loadable by the remoteproc driver. A delay was added to
the main method of core 0 in order to allow core 1 to
be loaded before attempting to trigger an interrupt. This
was necessary because core 1 is responsible for setting up
the interrupts. If core 0 didn't have a delay it would
send an interrupt before the interrupt configuration and
the example would hang.

Signed-off-by: Jason Reeder <jreeder@ti.com>