Pull the const table declarations into the headers
Previously each source file needed to duplicate the same
code in order to initialize a structure and tie it to one
of the constant table entries.

This code has been pulled up out of the source files into
the header files to remove confusing code from examples
as well as to reduce duplication.

Signed-off-by: Jason Reeder <jreeder@ti.com>