Update examples to use new header constant defs
The constant table structure definition was pulled into the
headers in the previous commit. This commit updates all of
examples to use the new headers and remove the duplication.

Also updated the RPMsg examples in this commit to clean up
the main.c file to make them more uniform between PRU0 and
PRU1. The deltas between the two examples were pulled up
into #define statements.

Signed-off-by: Jason Reeder <jreeder@ti.com>