Removing the usage of 16-bit bit fields
It was found that the assembly generated for 16-bit bit
fields may not be aligned on a word boundary. Because of
this, the bit field definitions were removed from the header
files and now their usage is being removed from these source
files.

If the memory accesses are not aligned on a correct boundary
then the writes may be ignored by the bus that you are
trying to update a peripheral register through. These can be
added back once the compiler/assembler is updated.

Signed-off-by: Jason Reeder <jreeder@ti.com>