Adding an RPMsg Echo example as lab 5
This example will first interact with the sample client
driver to show loopback with a Linux driver. A set of
instructions will be provided on a wiki that will walk
the user through modifying the example in order to use
the RPMsg_PRU linux driver that represents itself as a
character device in user space.

Signed-off-by: Jason Reeder <jreeder@ti.com>