Configure FCR.DMAMODE1=1 per TRM
From Brad's recommendation in the other Hardware UART example.

The TRM says "always write 1 to the DMAMODE1 bit".  It also mentions
in places that the DMA won't function properly without this setting.
Most likely this isn't critical for CPU driven operation, but as a
good practice we should always set this bit to 1.

Signed-off-by: Jason Reeder <jreeder@ti.com>