Correcting a stack overflow in the MAC example
The 'buf' array of operand structures was previously
instantiated locally in the main method. After stepping
through the code it was seen that this caused the result
variable to be pushed out into and address that does not
exist.

Moving the 'buf' array out of the main method corrected
this issue.

Signed-off-by: Jason Reeder <jreeder@ti.com>