Remove the divide by 4 from the const table example
There were two pointers to 32 bit arrays in the example. In
order to use a register offset in the array, the offset had
to be divided by 4 (4 bytes in 32 bits). This was confusing
so this commit changes the arrays from 32 bit arrays to 8
bit arrays. This removes the requirement for the divide by 4.

Signed-off-by: Jason Reeder <jreeder@ti.com>