Decouple the RPMsg library from the MBX internals
This commit updates the RPMsg library in order to remove
any dependencies on the MBX structure. Now, the examples
pass a 32 bit pointer to the message address that the
library needs write to in order to 'kick' the ARM.
Previously, the library had to include the mailbox
header file and use the instance variable provided. This
did not scale to the new devices, hence the change.

Signed-off-by: Jason Reeder <jreeder@ti.com>