Added a status check in the RPMsg examples
It was found that a race condition could occur during boot
where the virtio_rpmsg_bus module may not have been ready
for RPMsg operations. A check on a status bit was added to
the RPMsg PRU firmwares in order to wait under the Linux
modules were ready for communication.

Signed-off-by: Jason Reeder <jreeder@ti.com>