Refer to hdmi and sound node using absolute path
Previously, a patch was necessary to add a label to both
the hdmi and sound node in the am335x-boneblack.dts file in
order to be able to disable them in this dtsi file.

This commit moves away from using the labels to using the
absolute path of the nodes. This removes the dependency on
the label adding patch.

Signed-off-by: Jason Reeder <jreeder@ti.com>