Moving PRU_SHAREDMEM to PAGE 2 in the linker files
CCS can only access the PRU shared memory located at
subsystem address 0x10000 through PAGE 2. It was discovered
that if the DATA_SECTION pragma was used to place
initialized variables into the shared memory that CCS would
fail to load the program binary once it tried to validate
that the data was written correctly because it had not in
fact written any data to shared memory.

Once PRU_SHAREDMEM is moved to PAGE 2 in the command
linker file, CCS can successfully load and verify the
binary.

PAGE 1 versus PAGE 2 for the shared memory has no effect
on the pruss_remoteproc module's ability to load data
correctly into the shared memory.

Signed-off-by: Jason Reeder <jreeder@ti.com>