Update the software manifest to version 5.0.0
Due to changes in the underlying mechanisms of rpmsg, the
previous examples will no longer work with the Linux kernel
v4.4 drivers. For this reason the major revision number of
this software package is getting bumped to 5.0.0.

New rpmsg examples have been provided and a new software
manifest to cover this major revision has been added.

Signed-off-by: Jason Reeder <jreeder@ti.com>