Indentation and comment fixes in RPMsg code
Due to incorrect settings in one of the text editors being
used, there were indentation errors introduced throughout
the package. This commit cleans up the indentation in the
RPMsg example and library.

In order to reduce the development effort of updating
RPMsg code for multiple devices, a script was created to
generate the main.c file from a template. The previous
method of copy/paste/modify led to human error in the
comments that were not caught. The script corrected these
errors.

Signed-off-by: Jason Reeder <jreeder@ti.com>