Remove the hex utility step from lab projects
The output of the hex utility step was not being used in
the lab projects. This step has now been removed from the
two lab projects that were using it.

Examples of usage of the hex utility can still be found in
the pru_cape/pru_fw/ folder of the support package.

Signed-off-by: Jason Reeder <jreeder@ti.com>