Make it more clear that CCS is not needed to build fws
The previous comments and error messages in the Makefiles
of the support package could make it seem like Code
Composer Studio (CCS) was necessary to use the TI provided
PRU C compiler.

This is not the case as the PRU code gen tools (including
the C compiler/linker) can be downloaded separately from
CCS. This allows the firmwares in the support package to
be modified and rebuilt using only these Makefiles and
the PRU code gen tools.

Signed-off-by: Jason Reeder <jreeder@ti.com>