rsc_types: Sync up remoteproc resource types with TI 4.14 kernels
Update the default resource types to match the latest changes done on
the TI 4.14 Linux kernels. The previously unused RSC_INTMEM type is
re-purposed into a vendor pre-load custom resource RSC_PRELOAD_VENDOR
type. The RSC_CUSTOM is actually the post-load custom resource type,
RSC_POSTLOAD_VENDOR so has been renamed appropriately. The TYPE_CUSTOM
is still defined for now for compatibility purposes.

The fw_rsc_custom structure type is also updated to add provision for
the versioning of any custom resource type. The structure size itself
is not changed, but is split up to add the versioning using the higher
bytes within the previous sub_type element. Note that the name of this
structure is not updated to match the fw_rsc_vendor used on the Linux
kernel side.

Signed-off-by: Suman Anna <s-anna@ti.com>