pru_types: Deprecate fw_rsc_custom_ints version field
The first element of the fw_rsc_custom_ints structure has been
defined as a version field. The versioning of the custom resource
types has since been defined as a sub-field within the custom
resource type field itself in the fw_rsc_custom structure. So,
rename this version field as reserved. A value of 0 should be
continued to be used for this field for existing users. The field
itself is retained for now to maintain backward compatibility
of various firmwares, and will be removed later on.

Signed-off-by: Suman Anna <s-anna@ti.com>