pru_types: Add a new custom intc resource structure for ICSSG
The PRU INTC within ICSSG on K3 AM65x SoCs supports increased
number of host interrupts/channels and system events. The
current fw_rsc_custom_ints structure is not sufficient to
properly configure these additional host interrupts/channels
on the Linux side. So, a new custom interrupt resource type
structure fw_rsc_custom_ints_k3 is added to support this.
The Linux driver can support both versions simultaneously,
so the previous structure can still be used on AM65x resource
tables to save space if desired when the additional interrupt
configuration is not required. The new structure is mandatory
for the RTU cores though as their host interrupts are part
of the additional interrupts.

Note that this resource structure should be used with a version
value of 1 to allow the Linux driver to differentiate this new
structure from the previous resource structure.

Signed-off-by: Suman Anna <s-anna@ti.com>