examples: fixed dates
Some previous commits changed the copyright date from the original
year that the code was written to the current year (2018) on code
that was updated by the commit.

E.g., Copyright (c) 2015 Texas Instruments Incorporated
changed to
Copyright (c) 2018 Texas Instruments Incorporated

This commit fixes that mistake by retaining the original year the
code was written in, but adds the year that the code was last
modified.

Continuing the above example,
Copyright (c) 2015-2018 Texas Instruments Incorporated

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>